changed redirect logic location from Dashboard.tsx to PageContainer.tsx and test cases
